Add <cachedir> EXOSIMS parameter to config YAML file
EXOSIMS generates a large number of files and eventually causes a
"no space left on device" failure on the <sleec> machine after ~ 2 milliion
files are created.  It takes hours to remove the directory.
Letting the user to specify the cache directory is a workaround so that a
new run can commence while the previously filled directory is being removed.